+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add dialog box with image.
 * Add conda.yaml example.
 * Add .vscode folder.
 * Fix capturing exceptions and displaying them in the operator panel.